+++ packages/typescript-plugin/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @workflow/typescript-plugin
 
+## 4.0.1-beta.4
+
+### Patch Changes
+
+- 5078925: Add warning when invoking a workflow function without using `start()`
+- f973954: Update license to Apache 2.0
+- fcadd7b: Fix TypeScript LSP plugin for Node.js v20
+
 ## 4.0.1-beta.3
 
 ### Patch Changes