+++ packages/world-local/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @workflow/world-local
 
+## 4.0.1-beta.4
+
+### Patch Changes
+
+- 10309c3: Fix long-running steps to not time out after 5 minutes
+- f973954: Update license to Apache 2.0
+- Updated dependencies [f973954]
+  - @workflow/world@4.0.1-beta.4
+
 ## 4.0.1-beta.3
 
 ### Patch Changes